pyproject.toml
[tool.poetry]
name = "azfs"
version = "0.2.0"
description = ""
authors = ["yoshiki <yoshiki0911@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^1.0.3"
azure-identity = "^1.3.1"
azure-storage-file-datalake = "^12.0.0"
azure-storage-blob = "^12.3.0"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
pytest-mock = "^3.1.0"
azure-storage-queue = "^12.1.1"
sphinx = "^3.1.1"

[tool.poetry.dev-dependencies]
jupyterlab = "^2.1.1"
twine = "^3.1.1"
wheel = "^0.34.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
